renamed sample hello controller to FxmlController (for view purposes)
