+++ src/main/java/codes/pnk/HelloApplication.java
@@ -1,6 +1,6 @@
 package codes.pnk;
 
-import codes.pnk.controller.HelloController;
+import codes.pnk.view.FxmlController;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -12,7 +12,7 @@ public class HelloApplication extends Application {
     @Override
     public void start(Stage stage) throws IOException {
         FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("hello-view.fxml"));
-        HelloController controller = new HelloController(stage);
+        FxmlController controller = new FxmlController(stage);
         fxmlLoader.setController(controller);
 
         Scene scene = new Scene(fxmlLoader.load(), 320, 240);
