+++ src/main/java/module-info.java
@@ -7,6 +7,7 @@ module codes.pnk {
     opens codes.pnk to javafx.fxml;
 
     exports codes.pnk;
-    exports codes.pnk.controller;
     opens codes.pnk.controller to javafx.fxml;
+    exports codes.pnk.view;
+    opens codes.pnk.view to javafx.fxml;
 }
